fiber middleware which helps retrieving the upstream url refactored
