+++ internal/fiber/middleware/xfmphu/constants.go
@@ -1,7 +1,6 @@
-package xforwarded
+package xfmphu
 
 const (
 	xForwardedMethod = "X-Forwarded-Method"
-	xForwardedProto  = "X-Forwarded-Proto"
 	xForwardedURI    = "X-Forwarded-Uri"
 )
