+++ internal/fiber/middleware/xfmphu/handler.go
@@ -1,4 +1,4 @@
-package xforwarded
+package xfmphu
 
 import (
 	"context"
